describe command moved into vscode package
